install auxiliary libraries along with SURFEX binaries
When auxiliary libraries are built from source they are installed only in the
build directory. As result, when compiled SURFEX binaries are moved to their
final locations during the `install` step, they can't be run because of missing
shared libraries. To solve this problem, an additional step was introduced. This
step re-configures auxiliary libraries to be installed to the correct locations
before manually triggering `install` targets of those libraries.