add option to produce decorated names of executables
In the original build system for Open-SURFEX names of the produced
executables are decorated to reflect the used build configuration and
indicate whether user-provided Fortran sources were used or not.
This approach allows to build a number of different configurations of
Open-SURFEX and install them in the same directory without overwriting
of existing files. This commit introduces similar option that provides
decoration of Open-SURFEX executables based on the following schema:

<EXEC>-V<SURFEX VERSION>-<SOURCES>-<BUILD TYPE>-<COMPILER ID>[-OMP][-MPI]

where

EXEC           -- name of corresponding Open-SURFEX executable;
SURFEX VERSION -- version of the Open-SURFEX in form MAJOR-BUGFIX;
SOURCES        -- set to "MASTER" when no USER_SOURCES_DIR is provided, otherwise
                  set to the bottom directory of the USER_SOURCES_DIR path;
BUILD TYPE     -- set to the value of CMAKE_BUILD_TYPE variable;
COMPILER ID    -- Fortran compiler family as defined by cmake;
OMP            -- added when Open-SURFEX is built with OpenMP support;
MPI            -- added when Open-SURFEX is built with MPI support.