AppVeyor: use a pre-built cmake package
The cmake package in Cygwin repository has been updated, so now it
could be used to build Open-SURFEX. In Cygwin repository cmake-3.13.1-1
package is marked as a test version, so it is not clear how to install
it from the command line. Current solution is to manually download and
extract this package.